+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.4.32
 
-FEATURE: Resource count limits now include `frontends` to limit the number of frontends that are allowed to make connections to a share (https://github.com/openziti/zrok/issues/650)
+FEATURE: Resource count limits now include `share_frontends` to limit the number of frontends that are allowed to make connections to a share (https://github.com/openziti/zrok/issues/650)
 
 ## v0.4.31