+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 FEATURE: Resource count limits now include `share_frontends` to limit the number of frontends that are allowed to make connections to a share (https://github.com/openziti/zrok/issues/650)
 
+FIX: use controller config spec v4 in the Docker instance
+
 ## v0.4.31
 
 FEATURE: New "limits classes" limits implementation (https://github.com/openziti/zrok/issues/606). This new feature allows for extensive limits customization on a per-user basis, with fallback to the global defaults in the controller configuration.